Remove distributed coordination, replication, and routing modules
- Delete `internal/config/distributed.go` for distributed configuration management.
- Remove `internal/replication/distributed.go` for replication and failover handling.
- Eliminate `internal/routing/distributed.go` for query routing in distributed deployments.
