+++ internal/metako/app.go
@@ -1,6 +1,6 @@
-// Purpose    : Main application orchestration and lifecycle management
-// Context    : PostgreSQL replication management system core application
-// Constraints: Must handle graceful startup, shutdown, and component coordination
+// Purpose    : Main application for distributed pg-metako deployment
+// Context    : Manages distributed components with local node preference and coordination
+// Constraints: Must handle distributed coordination, failover, and routing seamlessly
 
 package metako
 
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"pg-metako/internal/config"
+	"pg-metako/internal/coordination"
 	"pg-metako/internal/database"
 	"pg-metako/internal/health"
 	"pg-metako/internal/logger"
@@ -20,16 +21,17 @@ import (
 	"pg-metako/internal/routing"
 )
 
-// Application represents the main application
+// Application represents the main distributed application
 type Application struct {
 	config             *config.Config
 	healthChecker      *health.HealthChecker
-	replicationManager *replication.ReplicationManager
-	queryRouter        *routing.QueryRouter
-	connectionManagers []*database.ConnectionManager
+	replicationManager *replication.DistributedManager
+	coordinationAPI    *coordination.CoordinationAPI
+	router             *routing.Router
+	localConnection    *database.ConnectionManager
 }
 
-// NewApplication creates a new application instance
+// NewApplication creates a new distributed application instance
 func NewApplication(cfg *config.Config) (*Application, error) {
 	// Validate configuration
 	if err := cfg.Validate(); err != nil {
@@ -39,52 +41,68 @@ func NewApplication(cfg *config.Config) (*Application, error) {
 	// Initialize health checker
 	healthChecker := health.NewHealthChecker(cfg.HealthCheck)
 
-	// Initialize replication manager
-	replicationManager := replication.NewReplicationManager(healthChecker)
-
-	// Initialize query router
-	queryRouter := routing.NewQueryRouter(replicationManager, cfg.LoadBalancer)
-
-	// Create connection managers for all nodes
-	var connectionManagers []*database.ConnectionManager
-	for _, nodeConfig := range cfg.Nodes {
-		connManager, err := database.NewConnectionManager(nodeConfig)
-		if err != nil {
-			return nil, fmt.Errorf("failed to create connection manager for node %s: %w", nodeConfig.Name, err)
-		}
-
-		// Add to replication manager (which also adds to health checker)
-		if err := replicationManager.AddNode(connManager); err != nil {
-			return nil, fmt.Errorf("failed to add node %s to replication manager: %w", nodeConfig.Name, err)
-		}
-
-		connectionManagers = append(connectionManagers, connManager)
-		logger.Printf("Added node %s (%s) to cluster", nodeConfig.Name, nodeConfig.Role)
+	// Create local database connection
+	localConnection, err := database.NewConnectionManager(cfg.LocalDB)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create local connection: %w", err)
 	}
+
+	// Initialize coordination API
+	coordinationAPI := coordination.NewCoordinationAPI(cfg)
+
+	// Initialize distributed replication manager
+	replicationManager := replication.NewDistributedManager(cfg, coordinationAPI, healthChecker, localConnection)
+
+	// Initialize unified router
+	router, err := routing.NewRouter(cfg, coordinationAPI)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create router: %w", err)
+	}
+
+	logger.Printf("Initialized distributed application for node: %s", cfg.Identity.NodeName)
+	logger.Printf("Local DB: %s (%s) at %s:%d", cfg.LocalDB.Name, cfg.LocalDB.Role, cfg.LocalDB.Host, cfg.LocalDB.Port)
+	logger.Printf("Cluster members: %d", len(cfg.ClusterMembers))
 
 	return &Application{
 		config:             cfg,
 		healthChecker:      healthChecker,
 		replicationManager: replicationManager,
-		queryRouter:        queryRouter,
-		connectionManagers: connectionManagers,
+		coordinationAPI:    coordinationAPI,
+		router:             router,
+		localConnection:    localConnection,
 	}, nil
 }
 
-// Start starts the application
+// Start starts the distributed application
 func (app *Application) Start(ctx context.Context) error {
-	logger.Println("Starting application components...")
+	logger.Println("Starting distributed application components...")
 
-	// Connect to all database nodes
-	for _, connManager := range app.connectionManagers {
-		connectCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-		if err := connManager.Connect(connectCtx); err != nil {
-			logger.Printf("Warning: Failed to connect to node %s: %v", connManager.NodeName(), err)
-		} else {
-			logger.Printf("Successfully connected to node %s", connManager.NodeName())
-		}
-		cancel()
+	// Connect to local database
+	connectCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	if err := app.localConnection.Connect(connectCtx); err != nil {
+		logger.Printf("Warning: Failed to connect to local database: %v", err)
+	} else {
+		logger.Printf("Successfully connected to local database: %s", app.config.LocalDB.Name)
 	}
+	cancel()
+
+	// Initialize router
+	if err := app.router.Initialize(ctx); err != nil {
+		return fmt.Errorf("failed to initialize router: %w", err)
+	}
+	logger.Println("Router initialized")
+
+	// Start coordination API
+	if err := app.coordinationAPI.Start(ctx); err != nil {
+		return fmt.Errorf("failed to start coordination API: %w", err)
+	}
+	logger.Println("Coordination API started")
+
+	// Start distributed replication manager
+	if err := app.replicationManager.Start(ctx); err != nil {
+		return fmt.Errorf("failed to start replication manager: %w", err)
+	}
+	logger.Println("Distributed replication manager started")
 
 	// Start health monitoring
 	if err := app.healthChecker.StartMonitoring(ctx); err != nil {
@@ -92,11 +110,10 @@ func (app *Application) Start(ctx context.Context) error {
 	}
 	logger.Println("Health monitoring started")
 
-	// Start failover monitoring
-	if err := app.replicationManager.StartFailoverMonitoring(ctx); err != nil {
-		return fmt.Errorf("failed to start failover monitoring: %w", err)
+	// Update remote connections
+	if err := app.router.UpdateRemoteConnections(ctx); err != nil {
+		logger.Printf("Warning: Failed to update remote connections: %v", err)
 	}
-	logger.Println("Failover monitoring started")
 
 	// Start a goroutine to periodically log cluster status
 	go app.statusReporter(ctx)
@@ -106,61 +123,76 @@ func (app *Application) Start(ctx context.Context) error {
 
 // Stop stops the application gracefully
 func (app *Application) Stop(ctx context.Context) error {
-	logger.Println("Stopping application components...")
+	logger.Println("Stopping distributed application components...")
 
-	// Stop monitoring
-	app.replicationManager.StopFailoverMonitoring()
-	logger.Println("Failover monitoring stopped")
+	// Stop distributed replication manager
+	if err := app.replicationManager.Stop(ctx); err != nil {
+		logger.Printf("Warning: Failed to stop replication manager: %v", err)
+	}
+	logger.Println("Distributed replication manager stopped")
 
+	// Stop coordination API
+	if err := app.coordinationAPI.Stop(ctx); err != nil {
+		logger.Printf("Warning: Failed to stop coordination API: %v", err)
+	}
+	logger.Println("Coordination API stopped")
+
+	// Stop health monitoring
 	app.healthChecker.StopMonitoring()
 	logger.Println("Health monitoring stopped")
 
-	// Close all database connections
-	for _, connManager := range app.connectionManagers {
-		if err := connManager.Close(); err != nil {
-			logger.Printf("Warning: Failed to close connection to node %s: %v", connManager.NodeName(), err)
-		} else {
-			logger.Printf("Closed connection to node %s", connManager.NodeName())
-		}
+	// Close router and all connections
+	if err := app.router.Close(); err != nil {
+		logger.Printf("Warning: Failed to close router: %v", err)
+	}
+	logger.Println("Router closed")
+
+	// Close local connection
+	if err := app.localConnection.Close(); err != nil {
+		logger.Printf("Warning: Failed to close local connection: %v", err)
+	} else {
+		logger.Printf("Closed local database connection")
 	}
 
 	return nil
 }
 
 // Run runs the application with signal handling
-func (app *Application) Run(ctx context.Context) error {
+func (app *Application) Run() error {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
 	// Start the application
 	if err := app.Start(ctx); err != nil {
 		return fmt.Errorf("failed to start application: %w", err)
 	}
 
-	logger.Println("Application started successfully")
-
-	// Wait for shutdown signal or context cancellation
+	// Setup signal handling
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
 
-	select {
-	case <-ctx.Done():
-		logger.Println("Context cancelled, stopping application...")
-	case <-sigChan:
-		logger.Println("Shutdown signal received, stopping application...")
+	logger.Println("Application started successfully. Press Ctrl+C to stop.")
+
+	// Wait for signal
+	<-sigChan
+	logger.Println("Received shutdown signal")
+
+	// Cancel context to stop all goroutines
+	cancel()
+
+	// Stop the application with timeout
+	stopCtx, stopCancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer stopCancel()
+
+	if err := app.Stop(stopCtx); err != nil {
+		return fmt.Errorf("failed to stop application gracefully: %w", err)
 	}
 
-	// Graceful shutdown
-	shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer shutdownCancel()
-
-	if err := app.Stop(shutdownCtx); err != nil {
-		logger.Printf("Error during shutdown: %v", err)
-		return err
-	}
-
-	logger.Println("Application stopped")
+	logger.Println("Application stopped successfully")
 	return nil
 }
 
-// statusReporter periodically reports cluster status
+// statusReporter periodically logs cluster status
 func (app *Application) statusReporter(ctx context.Context) {
 	ticker := time.NewTicker(30 * time.Second)
 	defer ticker.Stop()
@@ -175,46 +207,49 @@ func (app *Application) statusReporter(ctx context.Context) {
 	}
 }
 
-// logClusterStatus logs the current status of the cluster
+// logClusterStatus logs the current status of the distributed cluster
 func (app *Application) logClusterStatus() {
-	// Get current master
-	master := app.replicationManager.GetCurrentMaster()
-	if master != nil {
-		logger.Printf("Current master: %s", master.NodeName())
-	} else {
-		logger.Println("No current master available")
-	}
+	// Get current master from replication manager
+	currentMaster := app.replicationManager.GetCurrentMaster()
+	logger.Printf("Current master: %s", currentMaster)
 
-	// Get node statuses
-	statuses := app.replicationManager.GetAllNodeStatuses()
+	// Get cluster state from coordination API
+	clusterState := app.coordinationAPI.GetClusterState()
 	healthyCount := 0
-	for nodeName, status := range statuses {
+	for nodeName, status := range clusterState.Nodes {
 		if status.IsHealthy {
 			healthyCount++
 		}
-		logger.Printf("Node %s: healthy=%t, failures=%d, last_checked=%v",
-			nodeName, status.IsHealthy, status.FailureCount, status.LastChecked.Format(time.RFC3339))
+		logger.Printf("Node %s: healthy=%t, role=%s, last_heartbeat=%v",
+			nodeName, status.IsHealthy, status.Role, status.LastHeartbeat.Format(time.RFC3339))
 	}
 
-	// Get query router stats
-	stats := app.queryRouter.GetConnectionStats()
-	logger.Printf("Query stats: total=%d, reads=%d, writes=%d, failed=%d",
-		stats.TotalQueries, stats.ReadQueries, stats.WriteQueries, stats.FailedQueries)
+	// Get router stats
+	stats := app.router.GetStats()
+	logger.Printf("Router stats: total=%d, local=%d, remote=%d, reads=%d, writes=%d, failed=%d",
+		stats.TotalQueries, stats.LocalQueries, stats.RemoteQueries,
+		stats.ReadQueries, stats.WriteQueries, stats.FailedQueries)
 
-	logger.Printf("Cluster status: %d/%d nodes healthy", healthyCount, len(statuses))
+	logger.Printf("Cluster summary: %d/%d nodes healthy, local_preference=%.1f",
+		healthyCount, len(clusterState.Nodes), stats.LocalPreference)
 }
 
-// GetQueryRouter returns the query router for external use (e.g., HTTP API)
-func (app *Application) GetQueryRouter() *routing.QueryRouter {
-	return app.queryRouter
+// GetRouter returns the query router for external use
+func (app *Application) GetRouter() *routing.Router {
+	return app.router
 }
 
 // GetReplicationManager returns the replication manager for external use
-func (app *Application) GetReplicationManager() *replication.ReplicationManager {
+func (app *Application) GetReplicationManager() *replication.DistributedManager {
 	return app.replicationManager
 }
 
-// GetHealthChecker returns the health checker for external use
-func (app *Application) GetHealthChecker() *health.HealthChecker {
-	return app.healthChecker
+// GetCoordinationAPI returns the coordination API for external use
+func (app *Application) GetCoordinationAPI() *coordination.CoordinationAPI {
+	return app.coordinationAPI
+}
+
+// GetLocalConnection returns the local database connection for external use
+func (app *Application) GetLocalConnection() *database.ConnectionManager {
+	return app.localConnection
 }
